+++ src/QueryChangeDetector.js
@@ -10,14 +10,25 @@ class QueryChangeDetector {
     }
 
     /**
-     * @param {ChangeEvent}
-     * @return {{reexec: boolean, results: []}} if reexec==true, the query must rerun over the database
+     * @param {Object[]} resultsData of previous results
+     * @param {ChangeEvent[]} changeEvents
+     * @return {{mustReExec: boolean, results: []}} if reexec==true, the query must rerun over the database
      */
-    runChangeDetection(changeEvent) {
-        const docId = changeEvent.data.doc;
-        const previousResults = this.query.results;
-        const wasDocInResults = previousResults.find();
+    runChangeDetection(resultsData, changeEvents) {
+        if (changeEvents.length == 0) {
+            return {
+                mustReExec: false,
+                results: null
+            };
+        }
+//        const docId = changeEvent.data.doc;
+//        const previousResults = this.query.results;
+//        const wasDocInResults = previousResults.find();
         // TODO continue here
+        return {
+            mustReExec: true,
+            results: results
+        };
     }
 
 }
@@ -28,7 +39,7 @@ class QueryChangeDetector {
  * @param  {RxQuery} query [description]
  * @return {QueryChangeDetector}         [description]
  */
-export default function create(query) {
+export function create(query) {
     const ret = new QueryChangeDetector(query);
     return ret;
 }