+++ src/util.js
@@ -27,6 +27,8 @@ import 'rxjs/add/observable/from';
 import 'rxjs/add/observable/fromEvent';
 import 'rxjs/add/observable/merge';
 import 'rxjs/add/operator/publishReplay';
+import 'rxjs/add/observable/defer';
+
 
 import 'rxjs/add/operator/timeout';
 import 'rxjs/add/operator/delay';